+++ R/mappers.R
@@ -719,11 +719,6 @@ ibm_invalid_output.default <- function(mapper, input, state, ...) {
 ## inla.mesh ####
 
 #' @param mesh An `inla.mesh.1d` or `inla.mesh.2d` object to use as a mapper
-#' @param indexed logical; If `TRUE`, the `ibm_values()` output will be the
-#' integer indexing sequence for the latent variables. If `FALSE`, the knot
-#' locations are returned (useful as an interpolator for `rw2` models
-#' and similar).
-#' Default: `TRUE`
 #' @export
 #' @describeIn bru_mapper Creates a mapper for 2D `inla.mesh` objects
 bru_mapper.inla.mesh <- function(mesh, ...) {
@@ -2930,7 +2925,8 @@ ibm_n.bru_mapper_harmonics <- function(mapper, inla_f = FALSE, ...) {
 #' @export
 #' @rdname bru_mapper_methods
 ibm_jacobian.bru_mapper_harmonics <- function(mapper, input, state = NULL, inla_f = FALSE, ...) {
-  A <- Matrix::Matrix(0.0, NROW(input), ibm_n(mapper))
+  # Indexing into Matrix is slower than into matrix, so wait until the end to convert
+  A <- matrix(0.0, NROW(input), ibm_n(mapper))
   off <- 0
   if (mapper[["intercept"]]) {
     A[, 1] <- 1.0 * mapper[["scaling"]][1]
@@ -2940,10 +2936,11 @@ ibm_jacobian.bru_mapper_harmonics <- function(mapper, input, state = NULL, inla_
     input <- (input - mapper[["interval"]][1]) / diff(mapper[["interval"]])
     for (ord in seq_len(mapper[["order"]])) {
       scale <- mapper[["scaling"]][mapper[["intercept"]] + ord]
-      A[, off + 1] <- cos(2 * pi * input * ord) * scale
-      A[, off + 2] <- sin(2 * pi * input * ord) * scale
+      angle <- (2 * pi * ord) * input
+      A[, off + 1] <- cos(angle) * scale
+      A[, off + 2] <- sin(angle) * scale
       off <- off + 2
     }
   }
-  A
+  as(A, "Matrix")
 }